// src/models/interfaces/d365Language.ts
/* eslint-disable @typescript-eslint/naming-convention */
export interface ID365WebsiteLanguage {
	adx_name: string;
	adx_websitelanguageid: string;
	_adx_portallanguageid_value: string;
}

export interface ID365PortalLanguage {
	adx_displayname: string;
	adx_portallanguageid: string;
	adx_name: string;
	adx_languagecode: string;
}